+++ src/utilities/Utility.cpp
@@ -109,9 +109,8 @@ const string Utility::hex2str(const int number) {
 }
 
 const string Utility::removeChar(const string& str, char c) {
-	string result;
-	for (char currentChar : str)
-		if (currentChar != c) result += currentChar;
+	string result(str);
+	result.erase(std::remove(result.begin(), result.end(), c), result.end());
 	return result;
 }
 
@@ -123,13 +122,15 @@ const string Utility::extractChars(const string& str, char from, char to) {
 	return result;
 }
 
-const string Utility::getHomeDir() {
-	const char* homeDir = getenv("HOME");
-	if (homeDir) return string(homeDir);
+const std::string Utility::getHomeDir() {
+	struct passwd* pw = getpwuid(geteuid());
+	if (pw && pw->pw_dir) {
+		return std::string(pw->pw_dir);
+	}
 	return "/root";
 }
 
 const string Utility::getConfigDir() {
-	return string(getHomeDir() + "/.local/share/" PROJECT_NAME);
+	return string(getHomeDir() + "/.local/share/" PROJECT_NAME "/");
 }